Reviewer 1
The code seemed to run as intended, and the project popped as it should displaying two tabs at the top and different color bands on the main window.  The colored tea bands functioned properly, when I clicked on each one they prompted me to a separate page with a specific label for each one.  However, on the top buttons that displayed a lighter color, it was more difficult to read the text that they prompted when you clicked on them because the text is also white. The top button was white and then lighter shades of the brown followed, then leading to darker.  I could navigate between the two tabs easily, this functioned properly.  The two tabs could've been labeled as "Teas" and the other as "Camera" though to make it easier to understand.  I couldn't get the camera to work however, this may be due to the emulator I ran though. I don't have an external device used as an emulator with a camera function.  The code with the camera function seems correct and doesn't display any errors, but I'm not entirely sure if it is correct because I wasn't able to get it to function completely.  The code seemed to be correct with the book, and I couldn't identify any errors or missing semicolons.